feat #101: Dev용 로직 제거, Storyboard VC 제거
